feat: (frontend) change the like button to `p` when not logged in
